+++ packages/ionic-react/src/schematics/application/schema.d.ts
@@ -2,7 +2,6 @@ import { Linter } from '@nrwl/workspace';
 
 export interface ApplicationSchematicSchema {
   name: string;
-  skipFormat: boolean;
   directory?: string;
   tags?: string;
   unitTestRunner: 'jest' | 'none';