+++ packages/ionic-react/CHANGELOG.md
@@ -7,6 +7,7 @@
 - remove `classComponent` option from `application` schematic (now defaults to functional components)
 - remove `style` option from the `application` schematic (now defaults to CSS)
 - remove `pascalCaseComponent` option from the `application` schematic (now defaults to true)
+- remove `skipFormat` option from the `application` schematic (now defaults to false)
 
 # 4.1.0